Optimize ABLForcingAlgorithm for precursor simulations
- Move ABLForcingAlgorithm to `wind_energy` directory

- Convert ABLForcingAlgorithm to use the ABL Statistics object that already
  contains the spatially averaged velocities and temperatures as a function of
  height to determine driving pressure gradient

- Update ablNeutralEdge regression test to use the new ABL forcing class

- Tests on the OWEZ (9x9x1 km) precursor run indicates that the wall-clock time
  per Nalu timestep reduced to approximately 4s/time-step from 9s/time-step with
  the original implementation.
