+++ src/EnthalpyABLSrcNodeSuppAlg.C
@@ -5,7 +5,7 @@
 
 #include "EnthalpyABLSrcNodeSuppAlg.h"
 #include "Realm.h"
-#include "ABLForcingAlgorithm.h"
+#include "wind_energy/ABLForcingAlgorithm.h"
 
 // stk_mesh/base/fem
 #include <stk_mesh/base/Entity.hpp>
