+++ src/wind_energy/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_sources(GlobalSourceList
+  ABLForcingAlgorithm.C
   BdyHeightAlgorithm.C
   BdyLayerStatistics.C
   )
